+++ src/integrations/deepeval/__init__.py
@@ -1,0 +1,29 @@
+
+
+from src.integrations.deepeval.factory import create_deepeval_engine
+
+from src.integrations.deepeval.adapter import DeepEvalEngine
+
+from src.integrations.deepeval.exceptions import (
+    DeepEvalDependencyError,
+    DeepEvalExecutionError,
+    DeepEvalIntegrationError,
+    UnsupportedDeepEvalMetricError,
+)
+
+from src.integrations.deepeval.metrics import (
+    create_metric,
+    supported_metric_names,
+)
+
+__all__ = [
+    "DeepEvalDependencyError",
+
+    "DeepEvalEngine",
+    "DeepEvalExecutionError",
+    "DeepEvalIntegrationError",
+    "UnsupportedDeepEvalMetricError",
+    "create_metric",
+    "create_deepeval_engine",
+    "supported_metric_names",
+]